NF+RF: os.{chdir,getcwd} -> utils.{chpwd,getpwd}
The reason is that there is no bloody way to return
to original directory if original directory had
symlinked directories in its path.  The only way is to
cd PWD as in the environment.  BUT that one is set by the shell
and not but os.chdir.  So -- those functions become PWD
aware and adjust it upon chdir

This will be needed for carrying out unittests which would
need to change CWD/PWD to a specific directory