BF/RF: major -- direct logging to stderr by default
I don't see easier way to detect either we are running as a git custom special
remote backend or normally.  when as a backend we must not spit out anything
to stdout, which is dedicated to interactions with git-annex. BUT in debug loglevel
many things start to spit out logs (e.g. from config) before we get to the point
within the backend to e.g. redirect logging from stdout to stdin.  So easier would
be just to use stderr by default, as e.g. annex and wget do