No static interface imports anymore (closes: gh-408)
This doesn't implement a clever lazy-importing scheme, but it
nevertheless remove all static interface imports. In a nutshell:

Static imports are replaced with definitions of what needs importing,
plus an optional definition of target names for Python API and cmdline
interface.

This moves us into a position to make import more clever/leaner/lazier
if needed in the future.