NF: Proposal for datalad-native meta data format
Which invents as little as possible. Objectives:

1) Compact
2) Suitable for hand-editing (we have two JSON-based ones already)
3) Flexible enough to be able to capture pretty much anything I could
think of

Choosing Debian's RFC822 felt natural. Pretty much all these files are
composed by hand (that proof enough).

The idea is to merged the various ideas from Debian's metadata files
into a single spec file (control, copyright, ...). If necessary, we
could have multiple files of course, but it doesn't look like we need it
(yet).

At this point, the implementation represents the minimal technical
infrastructure to support this format on a demo level, and to get the
discussion going.

My goal is to implement support for the relevant parts from here:
https://wiki.debian.org/UpstreamMetadata#Fields

Main advantage would be stellar metadata support when generating Debian
packages for datalad datasets.

I don't think we need any more than that, and it is already somewhat
based on the DOAP vocabulary. At this point, however, I don't see the
need for YAML, and stuck to RFC822.

If we would ever want to, we could even adopt this as an extension:
https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
(not saying we need it!).