NF: Attempt to produce a standalone "modified" filter
This is aming to provide a generic replacement of the `--since`
functionality built in `publish`. The goal is to have `publish`
use this for faster upfront filtering, instead of the current
recursive self-invocation.

Moreover, we might want to consider using this in other places
too. Such as "`get` data for all files that were changed --since".