ENH: More flexible SSH remote calls
No quoting or lexing anymore, can now pass compound commands to a
remote machine -- without shoehorning it in. This is a precondition
to be able to prepare the execution environment for calling non-system
Git installations -- to come next.

The implementation still makes sure that the entire remote portion of
the command actually goes to the remote.