RF: Do not place default push config on publish anymore
Instead temporarily amend the push config to push the current branch,
or rather the underlying branch in case of managed branches.

We may want to have this behavior to be conditioned on an explicit
switch, or to only be active when there is no push config at all.

ATM it makes sense to me to do it like this.

This is work towards a resolution of gh-1275