RF+NF: Dataset.get_subdatasets() -> standalone `subdatasets`
In response to #1411

This implementation is a lot faster, as it completely avoids
instantiation of repositories, but wraps `git submodule` instead.
Moreover, it exposes more information in a more directly accessible way.

It uses the new command API, so all transformation and filter mechanisms
can be applied based on arbitrary subdataset properties.

The old `Dataset.get_subdatasets()` is left in place, as it is still
used a lot in the tests. All occurrences in the actual code base are
RF'ed. Use in test triggers a warning identifying the caller to
streamline further RF.

The new `subdatasets` command as advertised as a plumbing command in the
docs.

Tests in `test_subdatasets.py` have been extended to near 100% coverage,
only one safeguard line is missing.