RF: Dissolve _clone_from_any_source (fixes gh-1469)
This old helper provided an incomplete, alternative implementation
of what the `clone` command is doing now. This RF merged the aspects
of this function that are unique to the handling of submodules
into the _install_subds_from_flexible_source helper. A bunch of
assertions are added to make sure that helpers are only used in contexts
that they are intended for.