RF: Switch (back) `install` return values to be datasets
Adjust `install()` calls in its own test to assume the new default
behavior, but keep install calls in other tests to be more explicit.

But this time, only dataset instances for successful actions (and None
for unsuccessful attempts).

Also return a single items (instead of a one-item list) whenever
possible.

This does not yet have the result filter that exclude results that do
not match an input argument.