OPT+RF+DOC: Cheapo `sort_paths_into_subdatasets()`
For big dataset hierarchies the savings should be substantial,
as previously this function (called for any save operation)
would eventually ask for a full dataset tree, at least for
recursive operation.

This RF removes the need for subdataset requests completely,
and only traverses the filesystem for discovery (see previous
OPT), and not at all for sorting.