OPT: Second, faster subdataset query for recursion (fixes gh-1507)
The old one is still in available, because it is more flexible for the
price of making one Git call per dataset -- where the new one just makes
a single call in total. Switching is done automatically. The results are
the same, the speed is much improved for large trees.