RF: Major changes to `add`, now using annotated paths
This is a fairly radical change, as it relies on information in
annotations for making any decisions in `add`, instead of (re-)inpecting
the situation on the filesystem.

This should eventually make it a little faster, but more importantly
much easier to precisely define the criteria upon which to act, and
even more importantly be able to pass this accumulated information
down to the next command.

Which in this case will be `save` -- with its current problems on not
being able to decide what to do with which path.