RF+BK(somewhat): `save` using path annotation, save_ds_hierarchy obsolete
This RF makes use of the new stuff, and avoids a number of ambiguities
that previously arose from passing the content_by_ds dicts around. Now
we pass annotated paths, and annotate them with all the info we need.

This removes the save_dataset_hierarchy middle-man from the POV of
`save`. All commands could be RF'ed to use `save` directly now, and
go full generator even for this step.

This commits brings annotated paths at the door step of `save_dataset`
meaning that we can now address the issues that we are having with
untracked files. In this context this RF commit also brings the change
that we are no longer passing anything that was `add`ed directly and in
full to `save`. Instead we are hand-picking the subset of `add` results
that actually need saving and only pass those on -- as annotated files.

This commit leaves some tests broken. Some are annotated with reasoning
on why I think the tested behavior is actually undesired. Some others
are just TODO.