ENH: specify alternative sources for `clone`
This extends the already existing facility to generate candidate source
URL from a single argument to also be able to eat additional candidates
via an argument. This change enabled a tiny RF of `get` that allowed for
the removal of a loop of submodule URL candidates.

This PR also resolves the issue of intermediate error reporting
expressed in gh-1553, but not the other ideas in that issue.