ENH: Generate and dump valid JSON-LD context for search index fields
In theory this should enable us to resolve any key term in any search index
to its fundamental definition (given properly defined vocabularies in
all parser (which is not yet the case, and maybe never will be in
full)).

This context does not (yet) include defintions of controlled
vocabularies used in field values, unless such terms are explicitly
defined in a dataset's context already.