BF: ls -- Don't use sys.getdefaultencoding for unicode check
For Python 2, we should expect sys.getdefaultencoding to return
'ascii' regardless of what encoding the output device supports (e.g.,
see this stackoverflow answer [*]), so we shouldn't base our encoding
decision off of sys.getdefaultencoding.

Also, avoid a type error by explicitly handling a sys.stdout.encoding
value of None, which happens with Python 2 when stdout is piped.
We're disabling unicode in this case, which is consistent with the
next patch that will disable unicode if stdout isn't a tty.

[*] https://stackoverflow.com/questions/15530635/why-is-sys-getdefaultencoding-different-from-sys-stdout-encoding-and-how-does#33726891