RF: Move 'datalad run --rerun' to 'datalad rerun'
Otherwise, if we go in the direction of extending datalad run's rerun
capabilities, things will start to get a little crowded, with lots of
options only in effect if the --rerun flag is given.

This is almost entirely a code move, with no intention of functional
change.  The one exception is the behavior of 'datalad run' without a
command.  Before it was treated as a noop because the positional
argument had to be optional so that 'datalad run --rerun' would work.
Now the positional argument is required.

Closes #2072.