ENH: rerun: Add --onto option
This makes it possible to "replay" a sequence of commands from the
same point rather than just re-executing the commands at the tip of a
branch.  The latter is useful if you want to rerun commands with your
tree in a different state (e.g., a script modified), whereas with the
former you can compare the two tips (with, e.g., git log --left--right
--cherry-mark).

Note that the name "--onto" is intended to be reminiscent of git
rebase's --onto.

At this point --onto always enters a detached, but a --branch argument
will be introduced in an upcoming commit.