ENH: rerun: Cherry pick commits that are not in HEAD
Now that the start point can be specified with --onto, make it
possible to apply non-command commits on the new tip.

We may want to make this configurable via a command line option, but
for now, cherry pick the commit unless it is already included in HEAD.
We might also want to consider something like a `datalad rerun
--abort` option to help the user recover from bad rerun states such as
a cherry pick conflict.