ENH: rerun: Handle empty value for --onto as special case
This allows `datalad --onto=reva reva..revb` to be expressed as just
`datalad --onto= reva..revb`, which is useful when reva is a hash or
something like HEAD~3.  This value makes sense to treat as a special
case because it defines the common base for replaying a series of
commits.

Re: https://github.com/datalad/datalad/pull/2076#issuecomment-357826369