BF: rerun: Make --message override recorded message
The intention was for a message on the command line to override any
recorded message, which is useful for rerunning a single commit.  It's
probably never useful for rerunning a sequence of commits, so we might
want to add support for some kind template later.

Re: https://github.com/datalad/datalad/pull/2076#discussion_r163448394