BF: run: Don't try to run an empty command
Before the run/rerun command split, the command argument was allowed
to be empty to support `run --rerun` (and this case was properly
tested).  After the split, the tests were removed, but the argparse
spec wasn't updated to make the command required, so calling `datalad
run` with an empty command leads to an error.

Update the no command case to be a noop rather than making it required
because the next commit will restore --rerun for compatibility.