ENH: Rough implementation of old-style search inspired functionality
This is not meant to be a 1:1 reimplementation. From my POV it makes
little sense to implement things like multiple expression AND queries
when we have other search modes that provide a full query language.

That being said, given that we have to ability to tune each search mode
individually, and also add new ones, I do not have objections against
any further specialisations, as long as they to not add tones of
special-case options to the API.