ENH: addurls: Use all columns as metadata by default
The only exception is that, if the URL is contained in a single
column, it's excluded.

Note: I'd prefer to express the default "all columns" mode as
autometa=True rather than no_autometa=False, but I can't figure out
how to use the plugin interface to specify a false value.  Using
"autometa=False" on the command line sets autometa to the string
"False", which is of course true.  I've chosen to go with no_autometa
rather than testing `autometa == "False"`.