ENH: addurls: Add an argument for overriding empty strings
`extract`s signature is becoming even more unwieldy, but I'm not sure
what a better alternative is.  It could become a little more readable
with keyword arguments, but I'd rather not do that because it is just
repeating the values from dlplugin's signature.  Or I could put most
things in **kwds, but then it's hard to tell which arguments from
`dlplugin` are actually used in `extract`.  Or I could expose the
smaller helpers that `extract` wraps, but then there would no longer a
single entry point that takes care of all the extraction.  Instead all
callers (ATM, dlplugin and the tests) would have to repeat the same
sequence of function calls.