BF: export_to_figshare: Fix annex option by reversing it
A value of annex=False will still evaluate true because the string
"False" is passed.  The only way to get the false behavior is to pass
an empty string.  Fix this by reversing the meaning of the argument.
This way, no_annex=True is actually true (as the string "True") and
then the default value for the unspecified is False rather than the
string "False".

We're converting the plugins to commands, so the above fix isn't
really the motivation for this change.  Instead it's that, with the
new interface, this will allow a --no-annex switch to be passed to
disable the default behavior.

Fixes #2213.