BF: `configure` command (fixes gh-918)
First crude implementation of a command to configure DataLad and
datasets, interactively.

This is not meant as a replacement for the ConfigManager, nor as a
direct cmdline frontend for it. I believe `git config` works much better
for this case, and supports way more corner cases.

OTOH I think we do need something that bootstraps an initial
configuration for a fresh installation, and maybe for a new dataset
(until with have that templating mechanism).

There are no tests, and I won't write them, because I am not (yet)
confident that this a viable path to go down. In any case, the actual
interaction with the rest of the codebase is minimal (ui, and cfg).
And it already exposed a bug #2270.