ENH+RF: addurls: Follow git-annex's heuristic for extensions
Instead of having a "leftmost period" extension split, add fields that
are based on the behavior of git-annex's
splitShortExtensions (essentially, after splitting on extensions,
consider an element of the tail to be part of the extension if it and
all the following elements are fewer than 5 characters).

Put this behavior under the "*_root" and ""*_ext" fields, moving the
os.path.splitext behavior under "*_root_py" and "*_ext_py", because
the former seems more useful.