RF/BF: get_commit_date to return authored date by default, handle detached head as well
As pyout showed some times we were getting None for the date and it was
happening in detached head because we would just return None then since
there is no branch.

Later this function could be extended to print date of any commit, not just
the last one in the branch etc.

By default now returns "authored", not "committed" (thus rename) date since
it is the one of interest typically -- if date was altered, better be the
one specified