TST: run: Reorder commits to accommodate fixed values for GIT_*DATE
This test reruns a sequence of commits unto the same starting point.
It fails if GIT_AUTHOR_DATE and GIT_COMMITTER_DATE are set to fixed
values because the first commit in the sequence is a non-command
commit.  As such, all the information used to generate the commit ID,
including the parent commit, is identical to the original commit, so
the new branch doesn't diverge from the original one.

A command commit would not have the same problem because its message
changes when "chain" is modified, so reorder the sequence so that the
non-command commit comes after a command commit.