ENH: run: Support runs with non-zero exit codes
When the original --rerun flag was introduced (027c31d79), we started
aborting whenever the run command had an error because we don't to
add, e.g., a mistyped command to the log and then rerun it.

To allow the user to add a run commit even when there is a non-zero
exit, let's dump the log message to a file whenever a 'datalad run'
command has a non-zero code.  The user can then confirm that they want
to store the results of this command by running 'datalad add -F ...'.

Note: If the run command created subdatasets, the save command given
the in the message will only work after gh-2361.