NF: Add generic support for independent extension modules
In the spirit of #2326 this enables independent packages to provide
DataLad command (suites), simply by defining the required info in a
structure reachable via a setuptools entrypoint.

This following repo implements a rather minimal extension module that
demonstrates this. The idea is that people could fork this, if they
want to implement an extension.

https://github.com/datalad/datalad-module-template

If this gets installed, DataLad will offer an additional "hello..."
command in a dedicated (and labeled) suite.

Not sure how we want to test this, but we could start installing this
extension in the standard travis builds and see that it does what it
should.