ENH: create: Consider adding subdatasets when --nosave is given
If we're creating a subdataset and --nosave is given, we don't
register the subdataset with the superdataset because 'git submodule
add' doesn't make much sense if the subdataset doesn't have a commit
yet.  (It modifies the .gitmodules file but no gitlink entry is
created.)  As a consequence, 'datalad save -d. -r .' cannot be used to
save changes after creating a subdataset with --nosave.

But in most cases, there is already a commit in a subdatasets (the one
that sets the annex backend), so add the subdataset as long as it has
at least one commit.  If the subdataset is a regular git repository,
there are no commits and the previous behavior is retained.