RF: save: Move message_file parameter to the end
When I added message_file, I positioned it right after message in
__call__'s signature because they are closely related.

This, however, makes test_api.test_consistent_order_of_args fail.
That test checks that positional CLI arguments come first in the
signature.  This isn't true for save's message, but the test makes an
exception, arguing that it's save's most common argument.  This isn't
true for message_file, so re-position it.