ENH: Support fake commit dates
This is a minimal first step in supporting anonymized Git logs.  If
the datalad.dataset.fakedates configuration variable is enabled, all
commits made through datalad -- including annex commits -- should now
have faked dates.  (This commit might miss some datalad entry points
that create commits; we should check more extensively whether real
dates leak into the logs.)

The fake date is taken as one second after the latest commit date for
a commit in any local branch, or, if there are no commits yet, the
time is set to the same fixed date.  We most likely want to make it
possible to configure what the starting date is and what the method
for generating the fake date is.

Re: #1718