ENH: Add helper for verifying repo timestamps
When fake dates are enabled, there are two places where timestamps are
faked: 1) dates in the commit objects and 2) dates in the git-annex
branch blobs.  The check_dates function has the ability to extract
timestamps from both these sources and check whether they are newer or
older than a specified timestamp.

Right now there are no convenient entry points for this, but we could
expose it in a plugin that helps users check that they have not leaked
any dates into a repository created with --fake-dates.