ENH: Store datalad.fake-dates in .datalad/config
If we want to anonymize the dates of repository, it makes sense to
track the datalad.fake-dates config variable; otherwise the real date
could easily leak in from another location.  Up to this point, though,
we store it in .git/config because, if we set datalad.fake-dates *and*
commit .datalad/config in GitRepo.__init__, AnnexRepo.__init__ will
fail because it calls GitRepo.__init__ before it defines
AnnexRepo._batch, which AnnexRepo.precommit needs.

We could fix this by making sure AnnexRepo._batch is defined before
AnnexRepo calls GitRepo.__init__ or by making AnnexRepo.precommit
catch a NameError.  But all of this assumes we want to commit the
fake-dates configuration before we make any other commits.  If we
we're OK not doing that, no workaround is necessary.  'datalad create
--fake-dates' 1) sets datalad.fake-dates in .datalad/config, 2) sets
and commits the backend, and 3) makes the "new dataset" commit.  The
commit in step 3 includes the fake-dates setting (along with the
dataset ID and .datalad/.gitattributes).