BF: Treatment of a symlink pointing to a subdataset
We generally label paths of symlinks by what the symlink points to.
For example, a file in V5 annex, represented by a symlink, is labeled
'file' in path annotation.

Here I apply the same logic to a symlink that points to a subdataset.
We cannot call it a dataset (we have specific semnatics that would fail
otherwise, e.g. we cannot `add` such link as a submodule). So the only
option is to call it a 'directory'.