RF: @with_direct to run in one mode on travis/windows, but in both locally/buildbots
As a minimal refactoring of how those tests ran before testing both modes.
On Travis we have now dedicated matrix run to check direct, but not on
buildbots.  So I have decided to retain that sweeping.

On travis it might shave off 20-30 sec of time with this change since
would run only in "as config says" mode