BF: branch_blobs_in_tree: Handle empty ls-tree output
I encountered this in one of the crawl s3 tests where there is just a
single git-annex "branch created" commit.  I haven't dug into how to
reproduce this in a test.  In the normal case, there is ls-tree output
for any repo with a git-annex branch because 'git annex init' produces
two commit, one of them that points to the uuid.log.