RF: rerun: Show the commit message with "%B"
In the context of our recorded run commands, it produces the same
output as "%s%n%n%b" and is shorter.

In general, these two formats can produce different output if the user
doesn't put a blank line after the subject. %s then considers all the
lines as part of the subject and wraps them, whereas %B just spits out
the raw message.