RF: rerun: Retrieve message and extract run info in separate steps
We want to dump the full message in the --script output.  In order to
do that, we could to 1) modify get_commit_runinfo to return the full
message, 2) call git show twice to get the message, or 3) rework
get_commit_runinfo to take a commit message.  Go with 3, since it
avoids an unnecessary git call and allows this function to only be
concerned with extracting information from the commit message.