RF: rerun: Revert "message -> subject" renaming
In f0b082482 I argued that "subject" was a better way to refer to the
bits after "[DATALAD RUNCMD]" and before the json record.  But that
missed the point that these can include new lines, so they're not
restricted to being a "subject" (as in a one line Git commit message
subject).  Go back to calling them "message".