RF: repodates: Make guard against empty repositories more robust
It's better to handle the "no commits" issue within log_dates itself,
rather than check_dates, so that other callers of log_dates don't have
to worry about it.  Also change the check from "are there any local
branches?" because it's technically possible that there aren't while
some set of commits is still reachable from the specified
revisions (though that's not likely to happen in the wild).