BF: check_dates: Add kludge for compatibility with older dateutil
test_check_dates_invalid_date is failing in the
datalad-pr-docker-dl-nd80 buildbot.  Its traceback says that the error
is raised in /usr/lib/python2.7/dist-packages/dateutil/parser.py.
This means that the system has an older version of dateutil because
the parser.py file was removed before version 2.7.0 (in dateutil's
547b7a47b).

Recast this TypeError as a ValueError so that it matches the exception
raised by the most recent dateutil.