ENH: run: Add --output flag
While --input maps to 'git annex get', --output maps to 'git annex
unlock' or 'git rm', depending on whether content is present.  We
can't unconditionally unlock because that leads to an error if content
isn't present (gh-2432).

We're expanding the globs on the initial run in the same way we do
with --input.  As discussed in d5d00f227, this should be revisited
because there are cases for both --input and --output where the caller
would want to store the unexpanded patterns and re-glob on each rerun.

Re: #2432