BF: cli: Don't ignore the default result_filter
eval_results only uses the result_filter value from the interface
class if its kwargs argument doesn't include "result_filter" as a key.
This does the right thing for the Python API because the key is only
present if result_filter=None is explicitly specified.  For the CLI,
however, result_filter is always present, so the class's result_filter
is ignored.

To fix this, adjust call_from_parser to only add the result_filter key
if a filter was explicitly specified on the command line (via
--report-status or --report-status).

Fixes #2482.