RF+ENH: rerun: Process inputs before unlocking/removing any files
When content isn't present for a file that was added/modified in a
revision, it'd be better to remove the file rather than trying to
unlock it, which fails.  But if we do that, we should make sure we
process inputs first since that step may retrieve an added/modified
file (perhaps via --input=.).  To do both these things, piggyback on
run's outputs argument.