BF: run: Don't treat directories as globs
We want to be able to give a directory to --input.  Doing so allows a
command to depend on a potentially large number of files without
depending on everything with "." or globbing with --input='subdir/*',
which stores all the subdirectory's file names in the commit
message (which isn't pleasant to look at in the log and will currently
fail if the message reaches a certain length).

AnnexRepo.is_under_annex, however, fails when passed a directory
because it calls AnnexRepo.info, which raises an AssertionError.  It's
not clear what "under annex" should mean for a directory (contains one
or more annexed files?), but at the least AnnexRepo.info should
probably return information for the directory rather than raising an
AssertionError.

So avoid passing directories to is_under_annex.  If is_under_annex /
info are adjusted to handle directories, it might make sense to modify
_resolve_files to return a directory only when it contains an annexed
file.

Re: https://github.com/datalad/datalad/pull/2464#issuecomment-388535225