BF: handle DownloadErrors as well as CommandErrors while trying to enable special remotes
Necessary for compatibility with git-annex 6.20180510+gitg500f7ea78-1~ndall+1
which changed error message so we identify it as AccessFailedError

I am not sure if that is the correct solution here but the entire logic is not
clear for me -- i.e. why we would allow it to happen and just issue an info
message instead of failing.  If it is to catch cases where it is a git
repo on the other end -- there should be some analysis of the stderr