BF: run: Don't warn about automated outputs
When reruning, `outputs` is used to unlock the added/modified files.
But with --onto, these might not be present.  Don't issue a warning if
the glob doesn't match in this case because it's expected and the
output isn't a user-supplied output.

One issue with this solution is that we no longer warn if
user-supplied output globs don't match on rerun.  In order to do this,
we'd need to decouple automated and user-supplied outputs.