BF: run: Expand inputs/outputs from top-level when dataset is given
If a dataset is given on the command line (or when run is called as a
dataset method), expand globs relative to the top-level of the
dataset.  This follows the convention that datalad uses for path
handling.

The large test diff is due to an indentation change; we no longer need
a chpwd block.