NF: rerun: Add --report
Instead of iterating over the revisions and executing them
immediately, put revision/run information into result records.  This
allows us to have different handlers for these results.  At this
point, we have the standard handler that re-executes the commands and
the report handler that passes along the record to be displayed.

This is meant to be a minimal restructuring.  More information will be
reported results, and the --script functionality will also be moved to
into a handler.

Re: #2510