RF: rerun: Only check "skip or pick" once
We're operating on a revision range (with a single revision being
rev^..rev), so we can stop making the skip/pick decision once we've
hit the first pick commit.

This may avoid a few git calls, but the main advantage is that the
report results will now say whether it is a skip or a pick rather than
"cherry pick or skip".