ENH: cmdline: Disable abbreviations for long options
argparse's abbreviation feature makes it impossible to add a subparser
long option that is a prefix of a long option from the top-level
parser.  For example, adding --report to a subparser would fail with
"datalad: error: ambiguous option: --report could match
--report-status, --report-type".

Thanks to @yarikoptic for realizing that the abbreviation feature was
the root issue.

Re: https://github.com/datalad/datalad/pull/2544#issuecomment-391563279