ENH: run: Record dataset ID and check on rerun
When a command is run in a dataset, the same run information is
recorded in all subdatasets that have added/modified files.  Rerunning
the command from a subdataset isn't supported, so check the dataset ID
on rerun and, if it is different, skip or cherry pick the commit using
the same logic that's used for non-run commits.

Closes #1799.