RF: run: Always use a string for the command
From the command-line, the run command always comes in as a list.  If
it contains a single item, we unwrap it and pass the string directly
to the runner.  From the python api, the run command can come in as a
string or a list.  Commands in the string form are executed with
shell=True (so underneath with "/bin/sh -c"), which makes it possible
to use things like redirection and pipes in commands.

When we add support for commands with {inputs} and {outputs}
placeholders, the two representations will require separate
handling (e.g., if string, call format directly; if list, map format
call).  Instead of doing that, convert the list representation to the
string representation.