RF: run: Don't combine explicit and implicit outputs
During rerun, there are explicit outputs (those specified by --output
during the initial run) and implicit ones (those detected as
added/modified in the revision).

Keep these as separate streams because we are going to add support for
an outputs placeholder field in the command, and only explicit outputs
should be used there.