BF: run: Fix handling of subdirectory inputs/outputs
When executing within a dataset subdirectory, globbing is done
relative to that directory.  These paths are passed to the get and
unlock calls, which fail because these expect relative paths to be
relative to the dataset.

Adjust the paths to the appropriate value before calling get and
unlock.

That path handling inputs and outputs is getting unwieldy.  The next
commit will add a helper class.

Re: https://github.com/datalad/datalad/pull/2587#discussion_r191758806