NF: run: Add a placeholder for the run's working directory
This is useful when a command needs $PWD (e.g., mounting with
containers).  If it's hardcoded in the command, then that command
won't work reliably on different machines.  A more general approach is
to use a wrapper that resolves run-time values.  But many commands may
need $PWD, so let's add a placeholder to make a wrapper unnecessary in
this common, simple case.

Re: https://github.com/datalad/datalad-container/pull/35#issuecomment-393527318