RF: run: Unconditionally format the command
The command is only formatted if it looks like there's a placeholder
to be formatted.  The idea here was that it avoids unnecessary
work (particularly the sorting of the glob outputs).

However, this means that the behavior with respect to bracket escaping
differs depending on whether a placeholder is present.  That
distinction wouldn't be obvious to the caller, so it's better to
behave consistently here.  The extra sorting is unlikely to cause any
noticeable slowdown (and, if it does, we can deal with it then).

Thanks to @yarikoptic for raising this issue.