ENH+RF: run: Retain non-matching inputs and outputs
If globbing doesn't return any results, hang on to the original
pattern and let the callers handle the non-existing paths.  In the
case of inputs, warn when a path doesn't exist.  For outputs, log a
debug-level message rather than a warning because an output not
existing isn't a cause for alarm; it can happen when we rerun with
--onto.