BF: run: Fix resolution of relative pwd in symlinked directories
When a repository is under a symlinked directory, the relative path is
not calculated properly.  For example, in the TMPDIR="/var/tmp/sym
link" test case, the recorded run pwd would be something like
"../../var/tmp/sym link/repo/subdir".  Out tests didn't catch the
symlink issue despite explicitly testing the pwd value in
test_rerun_subdir because with_tempfile resolves the path with
realpath.

To trigger the failing tests, use with_tree instead because it doesn't
resolve the path.

To fix the recorded run pwd, pass pwd to get_dataset_root so that pwd
and dataset have the same leading path when they are fed to
os.path.relpath.

As a simple demonstration of the issue:

   import os
   from datalad.utils import getpwd
   print(os.path.relpath(getpwd(), "."))

When "/var/tmp/sym link" -> /tmp, this prints "." when called from
/tmp and "../var/tmp/sym link" when called from "sym link".