BF: Handle 401 errors from LORIS tokens
When a request is made with an incorrect username/password,
urllib.urlopen raises an HTTPError exception. Since datalad
is looking for AccessDeniedError exceptions in order to decide
whether to reprompt for a new token, it doesn't get caught
and there's no easy way to enter the correct credentials if
you don't get it right the first time.

This converts the exception to an AccessDeniedError so that
datalad asks if you want to re-enter your credentials.