ENH: cmdline: Add a newline to CommandError message
This is consistent with Python's rendering of exceptions (i.e., the
user doesn't include a newline in the message but the exception is
rendered with a trailing newline).

The main visible effect of this change is that, when a 'datalad run'
command fails, the message now ends with a newline.  In non-run cases,
particularly when Runner's log_online=False, this may result in two
newlines at the end of the message because Runner tacks the stderr to
the end of the message and that likely has a newline.  However, it's
preferable to have an extra newline rather than no newline because,
depending on the shell, the latter may disrupt the terminal's display
of the next prompt.

Re: https://github.com/datalad/datalad/pull/2446#issuecomment-384619844