DOC: Add summary of commands to datalad.api.__doc__
Make it easier for users to discover what methods are available by
looking at the help for the api module.  One concern with this change
could be that we're bringing in more modules than before, but
inspecting sys.modules suggests that this is not the case.

There is still some code duplication between api._command_summary and
main.setup_parser that could pulled out, but that can be done later if
someone thinks it's worth the effort.

Closes #2718.