BF: fixup handling of progress reports so they work also for addurl
tricky part is that addurl does not report a file for which it reports:
http://git-annex.branchable.com/todo/provide___39__file__39___in_--json-progress_record_for_addurl/
so code now relies on the fact that if it is only one expected item to be
dealt with then we do not need to match it by download_id which could be
defined by a key or a file.  Eventually this all might need to be RFed.

I also disabled total progress bar in case if there is only a single item
expected to be downloaded since it makes little to no sense then
(unless I missed why we had it that way... like if when we do not know
that there would be more than one