RF: rerun: Ignore multi-parent run commits
In order for these to occur, the user would have to do something like
'datalad run git merge topic'.  Don't support that because (1) it
would make it trickier to implement the logic for rerunning a revision
range with merges and (2) it's not clear that recording a merge as a
run command would ever be a sensible thing to do.