DOC: rerun: Redefine --onto=''
In preparation for teaching rerun to handle merges, we need to clarify
what --onto='' means.  The current definition is the value of --since.
Because we don't allow merges, that identifies an appropriate base.
But if the revision range is allowed to include merges, that could
give a less intuitive base.

For example, say we call 'datalad rerun --onto= --since=c d' with the
history below.

    d
    |
    M
    |\
    b c
    |/
    a
    |
    .

The revision list is 'c..d', so [b, M, d] would be replayed.  The
previous definition of --onto would identify c as the base to run b on
top of.

A rule that gives the same answer in the "no merges" case but a better
one in the "merges" case is "use the first parent of the first
revision with a run command".