RF: rerun: Make checks in rerun handler more explicit
The control flow for processing the results is confusing.  We use
dict.get to retrieve "rerun_action", so the reader would assume that
the key may not be present, but then there's an "else" body that
handles the result as rerun_action="run".

Instead, let's pass on any result that doesn't have a rerun_action.
The intention is clearer, and it make it possible for results that
don't have a have a run command to have a status other than "error",
which we'll use in the next commit.